+++ src/Adapter/AdapterInterface.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Aguimaraes\Adapter;
 
@@ -8,26 +8,26 @@ interface AdapterInterface
      * @param string $service
      * @param int $value
      */
-    public function setErrorCount($service = 'default', $value = 0);
+    public function setErrorCount(string $service, int $value);
 
     /**
      * @param string $service
      *
      * @return int
      */
-    public function getErrorCount($service = 'default');
+    public function getErrorCount(string $service): int;
 
     /**
      * @param string $service
      *
      * @return int
      */
-    public function getLastCheck($service = 'default');
+    public function getLastCheck(string $service): int;
 
     /**
      * @param string $service
      *
      * @return int
      */
-    public function updateLastCheck($service = 'default');
+    public function updateLastCheck(string $service): int;
 }