+++ src/Adapter/Dummy.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Aguimaraes\Adapter;
 
@@ -17,15 +17,15 @@ class Dummy implements AdapterInterface
     /**
      * @inheritdoc
      */
-    public function getErrorCount($service = 'default')
+    public function getErrorCount(string $service): int
     {
-        return isset($this->errorCount[$service]) ? $this->errorCount[$service] : 0;
+        return $this->errorCount[$service] ?? 0;
     }
 
     /**
      * @inheritdoc
      */
-    public function setErrorCount($service = 'default', $value = 0)
+    public function setErrorCount(string $service, int $value)
     {
         $this->errorCount[$service] = $value;
     }
@@ -33,15 +33,15 @@ class Dummy implements AdapterInterface
     /**
      * @inheritdoc
      */
-    public function getLastCheck($service = 'default')
+    public function getLastCheck(string $service):int
     {
-        return isset($this->lastCheck[$service]) ? $this->lastCheck[$service] : 0;
+        return $this->lastCheck[$service] ?? 0;
     }
 
     /**
      * @inheritdoc
      */
-    public function updateLastCheck($service = 'default')
+    public function updateLastCheck(string $service): int
     {
         return $this->lastCheck[$service] = time();
     }