+++ src/Adapter/Redis.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Aguimaraes\Adapter;
 
@@ -25,7 +25,7 @@ class Redis implements AdapterInterface
     /**
      * @inheritdoc
      */
-    public function setErrorCount($service = 'default', $value = 0)
+    public function setErrorCount(string $service, int $value)
     {
         $this->redis->set(
             sprintf('%s.%s.error_count', $this->prefix, $service),
@@ -36,7 +36,7 @@ class Redis implements AdapterInterface
     /**
      * @inheritdoc
      */
-    public function getErrorCount($service = 'default')
+    public function getErrorCount(string $service): int
     {
         return $this->getKey(
             sprintf('%s.%s.error_count', $this->prefix, $service)
@@ -46,7 +46,7 @@ class Redis implements AdapterInterface
     /**
      * @inheritdoc
      */
-    public function getLastCheck($service = 'default')
+    public function getLastCheck(string $service): int
     {
         return $this->getKey(
             sprintf('%s.%s.last_check', $this->prefix, $service)
@@ -56,12 +56,15 @@ class Redis implements AdapterInterface
     /**
      * @inheritdoc
      */
-    public function updateLastCheck($service = 'default')
+    public function updateLastCheck(string $service): int
     {
+        $updatedAt = time();
         $this->redis->set(
             sprintf('%s.%s.last_check', $this->prefix, $service),
-            time()
+            $updatedAt
         );
+
+        return $updatedAt;
     }
 
     /**
@@ -69,7 +72,7 @@ class Redis implements AdapterInterface
      *
      * @return int
      */
-    private function getKey($key)
+    private function getKey($key): int
     {
         if (!$this->redis->exists($key)) {
             return 0;