+++ src/CircuitBreakerInterface.php
@@ -1,8 +1,6 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Aguimaraes;
-
-use Aguimaraes\Adapter\AdapterInterface;
 
 interface CircuitBreakerInterface
 {
@@ -11,50 +9,15 @@ interface CircuitBreakerInterface
      *
      * @return bool
      */
-    public function isAvailable($service = 'default');
+    public function isAvailable(string $service): bool;
 
     /**
      * @param string $service
      */
-    public function reportFailure($service = 'default');
+    public function reportFailure(string $service);
 
     /**
      * @param string $service
      */
-    public function reportSuccess($service = 'default');
-
-    /**
-     * @param string $service
-     * @param int $default
-     *
-     * @return int
-     */
-    public function getThreshold($service = 'default', $default = 30);
-
-    /**
-     * @param int $value
-     * @param string $service
-     */
-    public function setThreshold($value, $service = 'default');
-
-    /**
-     * @param string $service
-     * @param int $default
-     *
-     * @return int
-     */
-    public function getTimeout($service = 'default', $default = 30);
-
-    /**
-     * @param int $value
-     * @param string $service
-     *
-     * @return mixed
-     */
-    public function setTimeout($value, $service = 'default');
-
-    /**
-     * @return AdapterInterface
-     */
-    public function getAdapter();
+    public function reportSuccess(string $service);
 }